STM32L4+: SDMMC is different from normal L4 series
The peripheral in the STM32L4+ (RM0432) is different from the peripheral
in some other STM32L4 series chips (RM0351).
It is still similar, but a significant amount of control/status bits
have changed or moved to a different bit offset.

The SVD files describe the older 'v1' peripheral while the datasheet and
cubeIDE describe the 'v2' version (very similar to the peripheral in
stm32h7 series). So I think the datasheet is correct and ST forgot to
update the peripheral in the SVD file..